rename : model 파일 삭제 및 domain 파일에 해당 로직 옯김
